+++ src/V1/Exceptions/UnsupportedType.php
@@ -43,31 +43,13 @@
 
 namespace GanbaroDigital\InstructionPipeline\V1\Exceptions;
 
-use GanbaroDigital\DIContainers\V1\FactoryList\Containers\FactoryListContainer;
+use GanbaroDigital\ExceptionHelpers\V1\BaseExceptions\UnsupportedType as BaseClass;
 
 /**
- * a FactoryList of our exceptions and how to create them
+ * exception thrown when we've been given a variable, and we do not know
+ * how to handle that PHP type
  */
-class InstructionPipelineExceptions extends FactoryListContainer
+class UnsupportedType extends BaseClass
 {
-    /**
-     * creates a FactoryList of exceptions and their factories
-     */
-    public function __construct()
-    {
-        // the exceptions that our library throws
-        $ourExceptions = [
-            'CannotFindInstructionBuilder::newFromInputParameter' => [ CannotFindInstructionBuilder::class, 'newFromInputParameter' ],
-            'CannotFindInstructionBuilder::newFromVar' => [ CannotFindInstructionBuilder::class, 'newFromVar' ],
-            'NotAnInstruction::newFromInputParameter' => [ NotAnInstruction::class, 'newFromInputParameter' ],
-            'NotAnInstruction::newFromVar' => [ NotAnInstruction::class, 'newFromVar' ],
-            'NotAnInstructionBuilder::newFromInputParameter' => [ NotAnInstructionBuilder::class, 'newFromInputParameter' ],
-            'NotAnInstructionBuilder::newFromVar' => [ NotAnInstructionBuilder::class, 'newFromVar' ],
-            'UnsupportedType::newFromInputParameter' => [ UnsupportedType::class, 'newFromInputParameter' ],
-            'UnsupportedType::newFromVar' => [ UnsupportedType::class, 'newFromVar' ],
-        ];
 
-        // build it
-        parent::__construct($ourExceptions);
-    }
 }